Fixed code error

+++ block_jacobi.py
@@ -62,7 +62,7 @@ def find_best_block_size(n, A, b, eval_key='run_time'):
     #plt.figure()
     #plt.plot(counter.residuals)
     #plt.show()
-    divisors = [i * n for i in FRACTION_CLASSES]
+    divisors = [int(i * n) for i in FRACTION_CLASSES]
     for divisor in divisors:
         # print(f"Testing block size: {divisor}")
         counter_pre = gmres_counter(disp=False)
